pyproject.toml
[project]
name = "warmup-semantic-segmentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lightning>=2.6.1",
    "optuna>=4.8.0",
    "optuna-integration[tensorboard]>=4.8.0",
    "setuptools==69.5.1",
    "tensorboard>=2.20.0",
    "tensorboard-plugin-profile>=2.22.0",
    "torchvision>=0.25.0",
]
